Extrais le gestionnaire de tirroir dans un module
